backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pydantic

google-genai
scikit-learn
joblib
pandas